% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.add2DB.R
\name{AFM.add2DB}
\alias{AFM.add2DB}
\title{Add AFM Images to the SQL Database}
\usage{
AFM.add2DB(baseSQLfile, IDs, verbose = TRUE)
}
\arguments{
\item{baseSQLfile}{path and file name of the SQLite database with the AFM images}

\item{IDs}{vector with list of unique file IDs to add}

\item{verbose}{logical, for verbose output}
}
\description{
Add AFM Images to the SQL Database
}
\author{
Thomas Gredig
}
