% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.create}
\alias{AFM.create}
\title{Creates a new AFMdata object from a height matrix}
\usage{
AFM.create(df, z.units = "nm", instrument = "reconstructed")
}
\description{
Creates a new AFMdata object from a height matrix
}
