% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.crop.R
\name{AFM.crop}
\alias{AFM.crop}
\title{Crops an AFM image}
\usage{
AFM.crop(
  obj,
  x0 = NA,
  y0 = NA,
  width.pixels = 1024,
  height.pixels = 1024,
  verbose = FALSE
)
}
\arguments{
\item{obj}{AFMdata object to be cropped}

\item{x0}{x origin location in pixels, if \code{NA} use GUI to click on origin}

\item{y0}{y origin location in pixels, if \code{NA} use GUI to click on origin}

\item{width.pixels}{width of cropped image in pixels}

\item{height.pixels}{height of cropped image in pixels}

\item{verbose}{output fitting parameters}
}
\value{
returns cropped image
}
\description{
Creates a subset of an AFM image by cropping the original
  image into a smaller AFM image object
}
\examples{
d <- AFM.artificialImage(type='calibration', width=100, height=100, verbose=FALSE)
plot(d)
d1 = AFM.crop(d, x0=45, y0=45, width.pixels = 50, height.pixels = 50)
plot(d1)

}
\author{
Thomas Gredig
}
