% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.flatten.R
\name{AFM.flattenLine}
\alias{AFM.flattenLine}
\title{Line Flatten Fit Parameters}
\usage{
AFM.flattenLine(
  obj,
  no = 1,
  skip = c(),
  region = NULL,
  tau_lower = 0.01,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{obj}{AFMdata image}

\item{no}{channel number}

\item{skip}{positions to skip (incorrect fit)}

\item{region}{dataframe with `lines` and `fit.px.lower` and `fit.px.upper` for each line}

\item{tau_lower}{percentage of data points to fit (1=100%)}

\item{verbose}{logical, output useful information if TRUE}
}
\value{
data frame with `m` (slope) and `b` (offset) for each line of the image
}
\description{
Line Flatten Fit Parameters
}
\author{
Thomas Gredig
}
