% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.getCSV.R
\name{AFM.getCSV}
\alias{AFM.getCSV}
\title{Converts AFM data to a 2D matrix}
\usage{
AFM.getCSV(obj, no = 1)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number}
}
\description{
Converts AFM data to a 2D matrix
}
\examples{
filename = AFM.getSampleImages(type='ibw')
d = AFM.import(filename)
m <- AFM.getCSV(d)
write.csv(m, "data.csv")
}
\author{
Thomas Gredig
}
