% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.getLine.R
\name{AFM.getLine}
\alias{AFM.getLine}
\title{Horizontal or Vertical Profile Line of AFM Image}
\usage{
AFM.getLine(
  obj,
  xPixel = NA,
  yPixel = NA,
  no = 1,
  dataOnly = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{obj}{AFMdata object}

\item{xPixel}{vertical line at this pixel (1 to image resolution), if \code{NA} will use yPixel}

\item{yPixel}{horizontal line at this pixel (1 to image resolution), if \code{NA} will use xPixel}

\item{no}{channel number}

\item{dataOnly}{if \code{TRUE}, returns data instead of AFMdata object}

\item{verbose}{if \code{TRUE}, output additional information}
}
\value{
AFMdata object with line data or data frame with data
}
\description{
create a profile data line for a particular
position given in pixels; either a horizontal or a vertical
line; this is also an easy way to reiterate through all lines of an image
}
\examples{
filename = AFM.getSampleImages('ibw')[1]
afmd = AFM.import(filename)
afmd2 = AFM.getLine(afmd, 50)
plot(afmd2, addLines = TRUE)
head(AFM.linePlot(afmd2, dataOnly = TRUE))


}
\seealso{
[AFM.lineProfile()]
}
\author{
Thomas Gredig
}
