% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.getSampleImages.R
\name{AFM.getSampleImages}
\alias{AFM.getSampleImages}
\title{Sample AFM image file names}
\usage{
AFM.getSampleImages(type = "*")
}
\arguments{
\item{type}{could be "ibw" for Asylum Research, "nid" for NanoSurf, 
"veeco" for Bruker/Veeco image,
"tiff" for Park, "force" for force spectroscopy, "resonance" for resonance curve,
if empty or "*", all files will be returned}
}
\value{
vector with path/filename to AFM sample images
}
\description{
returns sample AFM image files included with this library
}
\examples{
file.list = AFM.getSampleImages()
print(paste("Found",length(file.list),"sample files."))
}
\author{
Thomas Gredig
}
