% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.hhcf2.R
\name{AFM.hhcf2}
\alias{AFM.hhcf2}
\title{Height-Height Correlation Function 2}
\usage{
AFM.hhcf2(
  obj,
  no = 1,
  degRes = 100,
  numIterations = 1000,
  addFit = TRUE,
  r.percentage = 80,
  xi.percentage = 70,
  dataOnly = FALSE,
  fitparameters = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{Channel number}

\item{degRes}{resolution of angle, the higher the more, should be >100, 1000 is also good}

\item{numIterations}{Number of iterations (must be > 1000), but 1e6 recommended}

\item{addFit}{if \code{TRUE} a fit is added to the data}

\item{r.percentage}{a number from 10 to 100 representing the distance to compute, since the image is
square, there are not as many points that are separated by the full length, 80 is a good value, if there
is no fit, the value can be reduced to 70 or 60.}

\item{xi.percentage}{a number from 10 to 100 representing where correlation length could be found from maximum (used for fitting)}

\item{dataOnly}{if \code{TRUE} only return data frame, otherwise returns a graph}

\item{fitparameters}{if \code{TRUE} dataOnly will be set to \code{TRUE} and a list with data, fit, fit labels and fit parameters
will be returned.}

\item{verbose}{output time if \code{TRUE}}
}
\value{
Returns a graph with g(r) and $num indicating number of computations used for r. If dataOnly is TRUE then a data frame with only the
   x,y,z data is returned. If fitparameters is TRUE then a list that has the x,y,z data in the first element, the fitted curve is stored in the
   second element. The fit labels are stored in the third element, and the fit parameters (xi, sigma, H) are stored in the fourth element.
}
\description{
Changes the structure of the output variable for the AFM.hhcf function
from nanoAFMr package by Thomas Gredig et. al. Like its parent function, this
computes g(r) correlation function for an AFM
image with height information. The height information
must be encoded in a BMP file with a linear scale
Publication: http://iopscience.iop.org/article/10.1088/1742-6596/417/1/012069
Title: Height-Height Correlation Function to Determine Grain
       Size in Iron Phthalocyanine Thin Films
Authors: Thomas Gredig, Evan A. Silverstein, Matthew P Byrne
Amendments: Ryan Mizukami
Journal: J of Phys: Conf. Ser. Vol 417, p. 012069 (2013).
}
\examples{
filename = AFM.getSampleImages(type='tiff')
a = AFM.import(filename)
a = AFM.flatten(a)
r = AFM.hhcf2(a, numIterations = 5e5, fitparameters = TRUE)
head(r)                           # outputs HHCF data as a list

AFM.hhcf2(a, dataOnly = FALSE)    # outputs graph


}
\author{
Thomas Gredig, Ryan Mizukami
}
