% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.hhcf2QuickPlot.R
\name{AFM.hhcf2QuickPlot}
\alias{AFM.hhcf2QuickPlot}
\title{Height Height Correlation 2 Quick Plot}
\usage{
AFM.hhcf2QuickPlot(dataset, userLimitsx = NULL, userLimitsy = NULL)
}
\arguments{
\item{dataset}{a dataset (list) that was generated using AFM.hhcf2 and having fitparameters = \code{TRUE}}

\item{userLimitsx}{user defined axis limits for the x axis on a log scale. Default value is \code{NULL}}

\item{userLimitsy}{user defined axis limits for the y axis on a log scale. Default value is \code{NULL}}
}
\description{
Plot the data generated when using AFM.hhcf2 and having fitparameters = \code{TRUE}
Quick Note: userLimitsx and userLimitsy can both be ignored since ggplot will allow you to add additional properties to plots after first 
initialization of ggplot object. However, they are left in this function as arguments to allow for quick and clean plotting.
}
\examples{
AFM.hhcf2QuickPlot{r}

}
\author{
Ryan Mizukami
}
