% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.histogram.R
\name{AFM.histogram}
\alias{AFM.histogram}
\title{Histogram of AFM height data}
\usage{
AFM.histogram(obj, no = 1, binNo = 200, dataOnly = FALSE)
}
\arguments{
\item{obj}{An `AFMdata` object.}

\item{no}{Integer channel number.}

\item{binNo}{Integer number of bins to use.}

\item{dataOnly}{Logical; if `TRUE`, return histogram summary data instead of
a plot.}
}
\value{
A `ggplot` object when `dataOnly = FALSE`, otherwise a data frame
  with columns `mids` and `zDensity`.
}
\description{
Computes and optionally plots a histogram of one channel in an `AFMdata`
object. This is useful for checking whether a few outlier pixels dominate
the dynamic range before choosing `trimPeaks` or `setRange` in [plot()].
}
\examples{
img <- AFM.artificialImage(
  width = 64, height = 64, type = "gradient",
  addNoise = FALSE, verbose = FALSE
)
hist_data <- AFM.histogram(img, binNo = 40, dataOnly = TRUE)
head(hist_data, n = 6)
AFM.histogram(img, binNo = 40)


}
\author{
Thomas Gredig
}
