% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.history.R
\name{AFM.history}
\alias{AFM.history}
\title{History of AFM file generation}
\usage{
AFM.history(obj, verbose = FALSE)
}
\arguments{
\item{obj}{AFMdata object to be cropped}

\item{verbose}{output additional information if \code{TRUE}}
}
\value{
image history
}
\description{
outputs commands to recreate this particular AFM
  data image; if line profiles were added or the image was cropped
  that history will be shown; this allows reproducibility of the
  image.
}
\examples{
d <- AFM.import(AFM.getSampleImages('tiff'))
d <- AFM.lineProfile(d, 50,50,75,75, unitPixels = TRUE)
d <- AFM.lineProfile(d, 15,125,75,75, unitPixels = TRUE)
AFM.history(d)

}
\author{
Thomas Gredig
}
