% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.import}
\alias{AFM.import}
\title{Imports AFM file}
\usage{
AFM.import(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of AFM filename}

\item{verbose}{if \code{TRUE}, output additional information during loading of file}
}
\value{
AFMdata object
}
\description{
Use this function to create an AFMdata object from the filename;
four AFM formats (TIFF, NID, IBW, and 000) are supported. Use
\code{AFM.raster()} to create a data.frame from this object, or
use \code{plot()} to generate an image.
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw'))
summary(d)
plot(d)

}
\author{
Thomas Gredig
}
