% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.isImage}
\alias{AFM.isImage}
\title{checks if the object is an AFM image}
\usage{
AFM.isImage(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
\code{TRUE} if object is an AFM image, \code{FALSE} for frequency image
}
\description{
checks if the object is an AFM image
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw'))
AFM.isImage(d)
}
\author{
Thomas Gredig
}
