% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.linePlot.R
\name{AFM.linePlot}
\alias{AFM.linePlot}
\title{Plot AFM line profiles}
\usage{
AFM.linePlot(obj, no = 1, dataOnly = FALSE)
}
\arguments{
\item{obj}{An `AFMdata` object.}

\item{no}{Integer channel number.}

\item{dataOnly}{Logical; if `TRUE`, return the underlying data and do not
draw a plot.}
}
\value{
If `dataOnly = TRUE`, a data frame with columns `x`, `z`, and
  `type`; otherwise a `ggplot2` object.
}
\description{
Plots all profile lines stored in an `AFMdata` object for one channel.
Use [AFM.lineProfile()] or [AFM.getLine()] first to attach one or more lines.
}
\examples{
img <- AFM.artificialImage(
  width = 80, height = 80, type = "gradient",
  addNoise = FALSE, verbose = FALSE
)
img <- AFM.lineProfile(img, 0, 100, 1000, 100, unitPixels = FALSE)
img <- AFM.lineProfile(img, 0, 300, 1000, 300, unitPixels = FALSE)
AFM.linePlot(img)
head(AFM.linePlot(img, dataOnly = TRUE))
}
\seealso{
\code{\link{AFM.lineProfile}}
}
\author{
Thomas Gredig
}
