% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.lineProfile.R
\name{AFM.lineProfileMulti}
\alias{AFM.lineProfileMulti}
\title{Adds a Line Profile with Multiple Lines}
\usage{
AFM.lineProfileMulti(
  obj,
  x1 = NA,
  y1 = NA,
  x2 = NA,
  y2 = NA,
  N = 2,
  unitPixels = FALSE,
  verbose = FALSE
)
}
\description{
Adds a Line Profile with Multiple Lines
}
\seealso{
[AFM.lineProfile()]
}
