% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.particleSize.R
\name{AFM.particleSize}
\alias{AFM.particleSize}
\title{Find Particle Size in AFM image}
\usage{
AFM.particleSize(obj, floorHeight = 0.8, minParticleSize = 6)
}
\arguments{
\item{obj}{AFMdata object}

\item{floorHeight}{pixels with height below this threshold will be removed as background}

\item{minParticleSize}{minimum pixel size of a particle}
}
\value{
list with AFM image showing particles and also table with particle size
}
\description{
The image has to have particles that clearly contrast with the smooth
background.
}
