% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.readDB.R
\name{AFM.readDB}
\alias{AFM.readDB}
\title{Reads AFM object from SQL DB}
\usage{
AFM.readDB(mydb, ID, verbose = TRUE)
}
\arguments{
\item{mydb}{database connection from DBI pacakge}

\item{ID}{unique object file ID}

\item{verbose}{if \code{TRUE} outputs verbose comments}
}
\value{
S4 AFM object from nanoAFMr package
}
\description{
Reads AFM object from SQL DB
}
\examples{
fname = file.path(tempdir(), "afm.sqlite")
mydb <- DBI::dbConnect(RSQLite::SQLite(), fname)
afmFile = AFM.getSampleImages(type='tiff')[1]
a = AFM.import(afmFile)
AFM.writeDB(a, mydb, 45)
b = AFM.readDB(mydb, 45)
DBI::dbDisconnect(mydb)
plot(b)

}
\seealso{
[AFM.writeDB()]
}
