% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.readRatings.R
\name{AFM.readRatings}
\alias{AFM.readRatings}
\title{Reads AFM Ratings from SQLite Database}
\usage{
AFM.readRatings(dbFileName, verbose = FALSE)
}
\arguments{
\item{dbFileName}{full path and name of SQLite database}

\item{verbose}{if \code{TRUE} outputs verbose comments}
}
\value{
data frame with ratings to be saved
}
\description{
Returns AFM ratings for each AFM image and user.
}
\seealso{
[\code{\link{AFM.writeRatings}}]
}
\author{
Thomas Gredig
}
