% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.selectChannel.R
\name{AFM.selectChannel}
\alias{AFM.selectChannel}
\title{AFM select channels}
\usage{
AFM.selectChannel(obj, no)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number to select}
}
\value{
AFMdata object with only the selected channel
}
\description{
Allows to select a specific channel in order to reduce the size of the image
}
\seealso{
\code{\link{AFM.crop}}
}
\author{
Thomas Gredig
}
