% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.setLine.R
\name{AFM.setLine}
\alias{AFM.setLine}
\title{Set Pixels along Horizontal or Vertial Line of the AFM image}
\usage{
AFM.setLine(obj, lineData, xPixel = NA, yPixel = NA, no = 1, verbose = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{lineData}{vector with new values}

\item{xPixel}{vertical line at this pixel (1 to image resolution), if \code{NA} will use yPixel}

\item{yPixel}{horizontal line at this pixel (1 to image resolution), if \code{NA} will use xPixel}

\item{no}{channel number}

\item{verbose}{if \code{TRUE}, output additional information}
}
\value{
AFMdata object with new line
}
\description{
AFM.getLine returns the data from either a horizontal or vertial
line. The AFM.setLine sets the data for that line. Only a vector
with the data is submitted.
}
\examples{
a = AFM.artificialImage(height=10, width=10, verbose=FALSE)
b = AFM.getLine(a, yPixel = 1, dataOnly = TRUE)
lineData = rep(b$z[1], length(b$x))
a1 <- AFM.setLine(a, lineData, yPixel = 1)
plot(a1)


}
\seealso{
\code{\link{AFM.getLine}}
}
\author{
Thomas Gredig
}
