% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.specData.R
\name{AFM.specData}
\alias{AFM.specData}
\title{read spectroscopy channel data}
\usage{
AFM.specData(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
channels from spectroscopy header
}
\description{
read spectroscopy channel data
}
