% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.specHeader.R
\name{AFM.specHeader}
\alias{AFM.specHeader}
\title{read spectroscopy channel header info}
\usage{
AFM.specHeader(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
channels from spectroscopy header
}
\description{
read spectroscopy channel header info
}
