% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.to_CSV.R
\name{AFM.to_CSV}
\alias{AFM.to_CSV}
\title{Converts AFM data to a 2D matrix}
\usage{
AFM.to_CSV(obj, no = 1)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number}
}
\value{
Numeric matrix of `z` values with `y` rows and `x` columns.
}
\description{
Converts AFM data to a 2D matrix
}
\examples{
filename = AFM.getSampleImages(type='ibw')
d = AFM.import(filename)
m <- AFM.to_CSV(d)
write.csv(m, file.path(tempdir(), "data.csv"))
}
\author{
Thomas Gredig
}
