% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.writeDB.R
\name{AFM.writeDB}
\alias{AFM.writeDB}
\title{Write AFM Object to SQL Database}
\usage{
AFM.writeDB(obj, con, ID, vacuum = TRUE, verbose = FALSE)
}
\arguments{
\item{obj}{S4 AFMdata object, if object is \code{NULL}, then ID will be removed from database}

\item{con}{database connection from DBI package}

\item{ID}{unique object file ID}

\item{vacuum}{vacuums the database, if obj is \code{NULL} and something is deleted; this option saves space}

\item{verbose}{if \code{TRUE} outputs verbose comments}
}
\description{
Writes and AFM image to an SQL database; the ID is used to store the image. If an image
already exists with this ID, then it will be overwritten. If you use \code{NULL} as the
\code{obj}, then it will delete that ID. The space is only made available, if you vacuum
the database.
}
\examples{
sql_filename = tempfile(fileext = "sqlite")
con <- DBI::dbConnect(RSQLite::SQLite(), sql_filename)
AFM.writeDB(NULL, con, 0, vacuum=FALSE)
DBI::dbDisconnect(con)

}
\seealso{
\code{\link{AFM.readDB}}
}
\author{
Thomas Gredig
}
