% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.writeRatings.R
\name{AFM.writeRatings}
\alias{AFM.writeRatings}
\title{Write AFM Rating File to SQL Database}
\usage{
AFM.writeRatings(dbFileName, df_ratings, verbose = FALSE)
}
\arguments{
\item{dbFileName}{full path and name of SQLite database}

\item{df_ratings}{data frame with ratings to be saved}

\item{verbose}{if \code{TRUE} outputs verbose comments}
}
\description{
User rating information of specific AFM images is stored separately
in the AFM SQLite database.
}
\author{
Thomas Gredig
}
