% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFMdata}
\alias{AFMdata}
\title{AFM image class}
\usage{
AFMdata(
  data,
  x.conv,
  y.conv,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  history,
  description = "",
  fullFilename
)

AFMdata(
  data,
  x.conv,
  y.conv,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  history,
  description = "",
  fullFilename
)
}
\arguments{
\item{data}{list with objects ($z is a list with images, $freq is a resonance curve)}

\item{x.conv}{conversion factor from pixels to nm}

\item{y.conv}{conversion factor from pixels to nm}

\item{x.pixels}{number of pixels in x-direction}

\item{y.pixels}{number of pixels in y-direction}

\item{z.conv}{(not used)}

\item{z.units}{vector with units for $z (deg, m)}

\item{channel}{vector with names of channels}

\item{instrument}{name of instrument (Park, Cypher, NanoSurf, Veeco)}

\item{history}{history of file changes}

\item{description}{AFM image description or note}

\item{fullFilename}{name of file}
}
\description{
A S4 class to store and manipulate images from Atomic Force Microscopes.
It contains the names of channels (exact naming depends on the type/instrument)
The class supports multiple images; it can also contain additional information
such as lines, etc.
}
\section{Slots}{

\describe{
\item{\code{data}}{list with objects ($z is a list with images, $freq is a resonance curve)}

\item{\code{x.conv}}{conversion factor from pixels to nm}

\item{\code{y.conv}}{conversion factor from pixels to nm}

\item{\code{x.pixels}}{number of pixels in x-direction}

\item{\code{y.pixels}}{number of pixels in y-direction}

\item{\code{x.nm}}{length of image}

\item{\code{y.nm}}{height of image}

\item{\code{z.conv}}{(not used)}

\item{\code{z.units}}{vector with units for $z (deg, m)}

\item{\code{channel}}{vector with names of channels}

\item{\code{instrument}}{name of instrument (Park, Cypher, NanoSurf, Veeco)}

\item{\code{history}}{history of file changes}

\item{\code{description}}{AFM image description or note}

\item{\code{fullFilename}}{name of file}
}}

