% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMinfo.R
\name{AFMinfo.item}
\alias{AFMinfo.item}
\title{Specific information about AFM image}
\usage{
AFMinfo.item(obj, itemName = "")
}
\arguments{
\item{obj}{AFMinfo object}

\item{itemName}{name to retrieve (ScanRate, ScanAngle, ...)}
}
\value{
if \code{itemName} is empty, returns list of names, otherwise the value for the specific data item
}
\description{
Retrieve a specific information piece from the AFM image,
such as "ScanRate", "Description", etc. The item names depend
on the instrument; to find possible item names, leave
\code{itemName} empty.
}
\examples{
filename = AFM.getSampleImages(type='veeco')
h = AFMinfo(filename)
allNames = AFMinfo.item(h)
AFMinfo.item(h,"Description")
AFMinfo.item(h,"Operating mode")
print(allNames)
}
\author{
Thomas Gredig
}
