% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMinfo.R
\name{AFMinfo}
\alias{AFMinfo}
\title{AFMinfo object}
\usage{
AFMinfo(filename)
}
\arguments{
\item{filename}{full file name of instrument AFM file}
}
\value{
AFMinfo class object
}
\description{
header information about the AFM image, includes scan rate, widths, etc.
}
\examples{
filename = AFM.getSampleImages(type='veeco')
h = AFMinfo(filename)
summary(h)
}
\author{
Thomas Gredig
}
