% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.math.R
\name{get.Hsd}
\alias{get.Hsd}
\title{Computes the standard deviation in height Hsd}
\usage{
get.Hsd(z)
}
\arguments{
\item{z}{z component vector}
}
\description{
Computes the standard deviation in height Hsd
}
