% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{initialize,AFMdata-method}
\alias{initialize,AFMdata-method}
\title{Constructor method of AFMImage Class.}
\usage{
\S4method{initialize}{AFMdata}(
  .Object,
  data,
  x.conv,
  y.conv,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  history,
  description,
  fullFilename
)
}
\arguments{
\item{.Object}{an AFMdata object}

\item{data}{list with objects ($z is a list with images, $freq is a resonance curve)}

\item{x.conv}{conversion factor from pixels to nm}

\item{y.conv}{conversion factor from pixels to nm}

\item{x.pixels}{number of pixels in x-direction}

\item{y.pixels}{number of pixels in y-direction}

\item{z.conv}{(not used)}

\item{z.units}{vector with units for z (deg, m)}

\item{channel}{vector with names of channels}

\item{instrument}{name of instrument (Park, Cypher, NanoSurf, Veeco)}

\item{history}{history of file changes}

\item{description}{AFM image description or note}

\item{fullFilename}{name of file}
}
\value{
initialized AFMdata object
}
\description{
Constructor method of AFMImage Class.
}
