% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{print.AFMdata}
\alias{print.AFMdata}
\title{Print AFMdata object}
\usage{
\method{print}{AFMdata}(x, ...)
}
\arguments{
\item{x}{AFMdata object}

\item{...}{other arguments}
}
\value{
text with object information
}
\description{
prints essential information about the AFMdata object, which includes
the description, channel, image size, history, and filename
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw'))
print(d)
}
\author{
Thomas Gredig
}
