% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.math.R
\name{summary.AFMmath}
\alias{summary.AFMmath}
\title{Summary of computed AFM image parameters}
\usage{
\method{summary}{AFMmath}(object, ...)
}
\arguments{
\item{object}{AFMmath object}

\item{...}{other arguments}
}
\value{
prints a summary
}
\description{
Summary of computed AFM image parameters
}
\examples{
filename = AFM.getSampleImages(type='ibw')
summary(AFM.math.params(AFM.import(filename)))
}
