% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.add2DB.R
\name{AFM.add2DB}
\alias{AFM.add2DB}
\title{Add or Remove AFM Images to the SQL Database}
\usage{
AFM.add2DB(baseSQLfile, IDs, fIDfile = "data-raw/RAW-ID.csv", verbose = TRUE)
}
\arguments{
\item{baseSQLfile}{path and file name of the SQLite database with the AFM images}

\item{IDs}{vector with list of unique file IDs to add}

\item{fIDfile}{location of RAW-ID.csv file with path information for files}

\item{verbose}{logical, for verbose output}
}
\description{
If the ID is negative, then the image will be removed, any positive IDs are added.
If the IDs can be both positive and negative numbers, then images are both added and
removed; save any image, even partial AFM images.
}
\author{
Thomas Gredig
}
