% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.dataType}
\alias{AFM.dataType}
\title{Get AFMdata object type: image, frequency, force, spectroscopy}
\usage{
AFM.dataType(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
string with AFM type, "image", "noImage", "frequency", "spectroscopy"
}
\description{
Determine the type of the AFM file; this could be
    an `image` or `frequency` for resonance curve, or `spectroscopy` for
    a force vs distance spectroscopy file, `noImage` for a file 
    with data and no image
}
\examples{

# show the data type for each AFM file
fList = AFM.getSampleImages()
sapply(fList, function(x) { AFM.dataType(AFM.import(x)) })

}
\author{
Thomas Gredig
}
