% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.flatten.R
\name{AFM.flatten}
\alias{AFM.flatten}
\title{Flattens an AFM image using a plane fit}
\usage{
AFM.flatten(
  obj,
  no = 1,
  method = c("plane", "lineByLine", "slope"),
  slope = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number}

\item{method}{use the method to flatten the image:
\itemize{
   \item{"plane"}{`Default`: Fit a flat plane to the entire image and subtract}
   \item{"lineByLine"}{Fit each line and substract a linear fit}
   \item{"slope"}{Remove given slopes from each line, must provide `slope` parameter}
}}

\item{slope}{data.frame obtained from `AFM.flattenLine()`}

\item{verbose}{output fitting parameters}

\item{...}{additional arguments for method, such as tau_lower}
}
\value{
AFMdata object
}
\description{
uses the AFM.raster() function, makes a copy of the object
and fits a plane, returns the flattened object
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw'))
d2 = AFM.flatten(d)
plot(d2,graphType=2)

}
\seealso{
[AFM.flattenLine()]
}
\author{
Thomas Gredig
}
