% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.flatten.R
\name{AFM.flatten}
\alias{AFM.flatten}
\title{Flattens an AFM image using a plane fit}
\usage{
AFM.flatten(obj, no = 1, verbose = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number}

\item{verbose}{output fitting parameters}
}
\value{
flattened matrix with AFM image
}
\description{
uses the AFM.raster() function, makes a copy of the object
and fits a plane, returns the flattened object
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw'))
d2 = AFM.flatten(d)
plot(d2,graphType=2)
}
\author{
thomasgredig
}
