% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.histogram.R
\name{AFM.histogram}
\alias{AFM.histogram}
\title{Histogram of AFM image}
\usage{
AFM.histogram(obj, no = 1, binNo = 200, dataOnly = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number of the image}

\item{binNo}{number of bins in the histogram}

\item{dataOnly}{logical, if \code{TRUE} a data frame with the histogram
data is returned}
}
\value{
data frame or ggplot
}
\description{
Histogram for the AFMdata object is generated. You can set the number of bins
and either output an image or a data table.
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw'))
head(AFM.histogram(d, dataOnly=TRUE),n=20)
AFM.histogram(d)


}
\author{
Thomas Gredig
}
