% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.linePlot.R
\name{AFM.linePlot}
\alias{AFM.linePlot}
\title{Plots AFM line}
\usage{
AFM.linePlot(obj, no = 1, dataOnly = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number}

\item{dataOnly}{if \code{TRUE} no graph is returned}
}
\description{
Plots AFM line
}
\examples{
filename = AFM.getSampleImages(type='ibw')
d = AFM.import(filename)
AFM.lineProfile(d, 0,0, 2000,2000) -> d1
AFM.lineProfile(d1, 0,0, 100,2500) -> d2
AFM.linePlot(d2)
plot(d2,addLines=TRUE)
head(AFM.linePlot(d2, dataOnly=TRUE))
}
\seealso{
\code{\link{AFM.lineProfile}}, \code{\link{plot.AFMdata}}
}
\author{
Thomas Gredig
}
