% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.math.R
\name{AFM.math.params}
\alias{AFM.math.params}
\title{Various computed AFM image parameters}
\usage{
AFM.math.params(obj)
}
\arguments{
\item{obj}{AFMdata class}
}
\value{
structure with various computed AFM image parameters, such
as roughness
}
\description{
Various computed AFM image parameters
}
\examples{
filename = AFM.getSampleImages(type='ibw')
AFM.math.params(AFM.import(filename))
}
