% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.partial.R
\name{AFM.partial}
\alias{AFM.partial}
\title{Identifies Partial AFM Image}
\usage{
AFM.partial(afmd)
}
\arguments{
\item{afmd}{AFMdata object}
}
\description{
Sometimes an AFM image is recorded partially, this function
returns \code{TRUE}, if the image is a partial AFM image.
}
