% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.rate.R
\name{AFM.rate}
\alias{AFM.rate}
\title{Rate AFM Images}
\usage{
AFM.rate(dbFileName, IDs = NA, verbose = FALSE)
}
\arguments{
\item{dbFileName}{full path and name of SQLite database}

\item{IDs}{vector with AFM file IDs}

\item{verbose}{if TRUE additional information is output}
}
\description{
The images are displayed and rated by a user; if the user has rated
an image already, it will not show up anymore, if the image has not
yet been rated by that user, it will show and the user can enter a

rating usually: `high`, `ok`, `low`
}
