% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.writeDB.R
\name{AFM.writeDB}
\alias{AFM.writeDB}
\title{Write AFM object to SQL DB}
\usage{
AFM.writeDB(obj, mydb, ID, verbose = TRUE)
}
\arguments{
\item{obj}{S4 AFM object from nanoAFMr package}

\item{mydb}{database connection from DBI package}

\item{ID}{unique object file ID}

\item{verbose}{if \code{TRUE} outputs verbose comments}
}
\description{
AFM S4 image is written to an SQL database in the form
of a data.frame; the ID will save it in a particular table
in the database, if another AFM image with the same ID
exists, it will be overwritten.
}
\seealso{
[AFM.readDB()]
}
