% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.AFMdata}
\alias{plot.AFMdata}
\title{Graph Atomic Force Microscopy Image}
\usage{
\method{plot}{AFMdata}(
  x,
  no = 1,
  mpt = NA,
  graphType = 1,
  trimPeaks = 0.01,
  fillOption = "viridis",
  addLines = FALSE,
  redBlue = FALSE,
  verbose = FALSE,
  quiet = FALSE,
  setRange = c(0, 0),
  ...
)
}
\arguments{
\item{x}{AFMdata object}

\item{no}{channel number of the image}

\item{mpt}{midpoint for coloring}

\item{graphType}{1 = graph with legend outside, 2 = square graph with line bar, 3 = plain graph, 4 = plain graph with scale, 5 = legend only}

\item{trimPeaks}{value from 0 to 1, where 0=trim 0\% and 1=trim 100\% of data points, generally a value less than 0.01 is useful to elevate the contrast of the image}

\item{fillOption}{can be one of 8 color palettes, use "A" ... "H", see scale_fill_viridis}

\item{addLines}{if \code{TRUE} lines from obj are added to graph, lines can be added with  AFM.lineProfile for example}

\item{redBlue}{if \code{TRUE} output red / blue color scheme}

\item{verbose}{if \code{TRUE} it outputs additional information.}

\item{quiet}{if \code{TRUE} then no output at all}

\item{setRange}{vector with two values, such as c(-30,30) to make the scale from -30 to +30}

\item{...}{other arguments, such as col='white' to change color of bar}
}
\value{
ggplot graph
}
\description{
Graphs an atomic force microscopy image with or without legends depending on the graphing type.
An AFM image could include multiple channels (topography, phase, etc.), use \code{summary} to 
view all channels and the submit the number of the channel. Generally, the image peaks and valleys
are removed to increas contrast; for example a trimPeak of 0.01 would remove 1 percent of the
height data points as determined from the histogram. Additionally, the fillOption and redBlue
can add different colors. Since the returning object is a ggplot graph, it is possible to add
themes and additional aesthetics.
}
\author{
Thomas Gredig
}
